Refactored the Feedback app to contain all rest code in the "rest_framework" folder
