+++ api/app/signals/apps/api/urls.py
@@ -28,7 +28,7 @@ from signals.apps.api.views import (
     StatusMessageTemplatesViewSet,
     StoredSignalFilterViewSet
 )
-from signals.apps.feedback.views import FeedbackViewSet, StandardAnswerViewSet
+from signals.apps.feedback.rest_framework.views import FeedbackViewSet, StandardAnswerViewSet
 from signals.apps.search.rest_framework.views import SearchView
 from signals.apps.users.rest_framework.views import (
     AutocompleteUsernameListView,
