+++ api/app/signals/apps/feedback/rest_framework/exceptions.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from rest_framework.exceptions import APIException
 from rest_framework.status import HTTP_410_GONE
