+++ api/app/signals/apps/feedback/rest_framework/serializers.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from django.utils import timezone
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
